+++ tox.ini
@@ -13,7 +13,7 @@ deps =
 
 commands=
     flake8 releasetasks
-    py.test --cov=releasetasks --cov-report term-missing --doctest-modules releasetasks
+    py.test --verbose --cov=releasetasks --cov-report term-missing --doctest-modules releasetasks
 
 [testenv:py27-coveralls]
 deps=
